Add URL Params in Examples
